+++ package.json
@@ -123,8 +123,6 @@
     "optimizer.cjs",
     "optimizer.mjs",
     "optimizer.d.ts",
-    "prefetch.js",
-    "prefetch.debug.js",
     "qwikloader.js",
     "qwikloader.debug.js",
     "qwikloader.optimize.js",
