+++ scripts/build.ts
@@ -15,7 +15,6 @@ import {
 import { submoduleCore } from './submodule-core';
 import { submoduleJsxRuntime } from './submodule-jsx-runtime';
 import { submoduleOptimizer } from './submodule-optimizer';
-import { submodulePrefetch } from './submodule-prefetch';
 import { submoduleQwikLoader } from './submodule-qwikloader';
 import { submoduleServer } from './submodule-server';
 import { submoduleTesting } from './submodule-testing';
@@ -70,7 +69,6 @@ export async function build(config: BuildConfig) {
         submoduleJsxRuntime(config),
         submoduleQwikLoader(config),
         submoduleBuild(config),
-        submodulePrefetch(config),
         submoduleTesting(config),
         copyFiles(config),
       ]);
