--- src/prefetch-worker.ts
@@ -1,2 +1,0 @@
-import { setUpWebWorker } from './bootloader-shared';
-setUpWebWorker(self, fetch);
