+++ src/qwikloader-entry.ts
@@ -1,0 +1,3 @@
+import { qwikLoader } from './qwikloader';
+
+qwikLoader(document);
