+++ src/server/components.ts
@@ -1,5 +1,5 @@
 import { FunctionComponent, jsx } from '@builder.io/qwik';
-import { getQwikLoaderScript, getQwikPrefetchScript } from './scripts';
+import { getQwikLoaderScript } from './scripts';
 
 /**
  * @alpha
@@ -16,24 +16,6 @@ export const QwikLoader: FunctionComponent<QwikLoaderProps> = ({ events, debug }
   return jsx('script', {
     // It is tempting to add `type="module"` but that breaks in FF because it allows `scroll`
     // event to fire before the module is loaded.
-    // type: 'module',
     children: [getQwikLoaderScript({ events, debug })],
-  });
-};
-
-/**
- * @alpha
- */
-export interface QwikPrefetchProps {
-  debug?: boolean;
-}
-
-/**
- * @alpha
- */
-export const QwikPrefetch: FunctionComponent<QwikPrefetchProps> = ({ debug }) => {
-  return jsx('script', {
-    type: 'module',
-    children: [getQwikPrefetchScript({ debug })],
   });
 };
